// MtMInterviewApp/src/type.ts
export type DataType = {
  ID: string;
  Category: string;
  Caption: string;
  Company: string;
  StartDate: string;
  EndDate: string;
  TicketType: string;
  TicketPrice: string;
  TicketPurchaseLink: string;
  ContactPerson: string;
  ContactTel: string;
  Introduction: string;
  WebsiteLink: string;
  YoutubeLink: string;
  ImageFile: string;
  CreateDate: string;
  Venue: string;
  SessionStartDate: string;
  SessionEndDate: string;
  City: string;
  Area: string;
  Address: string;
  Longitude: number;
  Latitude: number;
  RelatedLink: string;
};

export const initDataType = {
  ID: '',
  Category: '',
  Caption: '',
  Company: '',
  StartDate: '',
  EndDate: '',
  TicketType: '',
  TicketPrice: '',
  TicketPurchaseLink: '',
  ContactPerson: '',
  ContactTel: '',
  Introduction: '',
  WebsiteLink: '',
  YoutubeLink: '',
  ImageFile: '',
  CreateDate: '',
  Venue: '',
  SessionStartDate: '',
  SessionEndDate: '',
  City: '',
  Area: '',
  Address: '',
  Longitude: 0,
  Latitude: 0,
  RelatedLink: '',
};